New batch of HTTP/2 protocol fixes (#2425)
* batch of fix

* revert example change

* more fixed

* even more fixes

* treat window increment size of 0 as error

* better trailer check

* trailer header filters

* format

* update all stream window size one first initial window size update

* format

* fix redis on Windows

* format

* fix again

* format

* better performance on download

* windows fix

* format

* set higher default window size

* format

* show error on protocol error

* remove unused function

* optimize queuing body being sent

* format

* add support for http/2 client to send ping

* format

* support ending ping frames

* avoid UAF

* respect server table size

* increase max encoded table size to 2MB

* fixes

* avoid UAF haneling errored stream

* upper case header treated as error

* fix typo

* treat header errors as protocol error

* RFC compliant GOAWAY ending

* graceful stream level error closing

* proper check if encrypted

* relax data frame error to stream level

* DROGON_DISABLE_HTTP2_CLIENT to disable HTTP/2 in client

* update changelog

* set version

+++ nosql_lib/redis/src/RedisConnection.cc
@@ -70,7 +70,9 @@ void RedisConnection::startConnectionInLoop()
     redisContext_->ev.cleanup = cleanup;
     redisContext_->ev.data = this;
 
-    channel_ = std::make_unique<trantor::Channel>(loop_, redisContext_->c.fd);
+    channel_ = std::make_unique<trantor::Channel>(loop_,
+                                                  static_cast<int>(
+                                                      redisContext_->c.fd));
     channel_->setReadCallback([this]() { handleRedisRead(); });
     channel_->setWriteCallback([this]() { handleRedisWrite(); });
     redisAsyncSetConnectCallback(
